requirements.txt
fastapi==0.65.1
uvicorn==0.14.0
motor==3.5.1
bcrypt==4.2.0
python-jose==3.3.0



pip list


Package            Version
------------------ ---------
annotated-types    0.7.0
anyio              4.4.0
bcrypt             4.2.0
certifi            2024.8.30
charset-normalizer 3.3.2
click              8.1.7
colorama           0.4.6
distlib            0.3.8
dnspython          2.6.1
ecdsa              0.19.0
fastapi            0.114.0
filelock           3.15.4
h11                0.14.0
idna               3.8
lxml               5.3.0
markdown-it-py     3.0.0
mdurl              0.1.2
motor              3.5.1
pip                24.2
pipenv             2024.0.1
platformdirs       4.2.2
pyasn1             0.6.1
pydantic           2.9.1
pydantic_core      2.23.3
Pygments           2.18.0
pymongo            4.8.0
python-jose        3.3.0
requests           2.32.3
rich               13.8.1
rsa                4.9
setuptools         74.1.2
six                1.16.0
sniffio            1.3.1
starlette          0.38.5
typing_extensions  4.12.2
urllib3            2.2.2
uvicorn            0.30.6
virtualenv         20.26.3